refactored Course component as a separate module
